feat(media): add S3 presigned URL support
- Update Prisma schema and Go types to use BigInt/Int64 for media size.
- Add utils/s3.go to generate presigned PUT URLs for S3 objects.
- Modify MediaService to generate and include media URL.
- Introduce .gvmrc and docker/.gitignore for project configuration.

+++ backends/design/types/media.go
@@ -7,7 +7,7 @@ import (
 var MediaInput = Type("MediaInput", func() {
 	Attribute("filename", String)
 	Attribute("mimeType", String)
-	Attribute("size", Int)
+	Attribute("size", Int64)
 	Attribute("bucket", String)
 	Attribute("key", String)
 
@@ -23,7 +23,7 @@ var Media = ResultType("application/vnd.media+json", func() {
 		Attribute("url", String)
 		Attribute("filename", String)
 		Attribute("mimeType", String)
-		Attribute("size", Int)
+		Attribute("size", Int64)
 		Attribute("bucket", String)
 		Attribute("key", String)
 		Required("id", "mediaType", "url")
